Complicator
Added Complicator
Added 10th ShowerRunner

+++ Assets/Scripts/Managers/SaveLoadSceneData.cs
@@ -7,5 +7,5 @@ using UnityEngine;
 [Serializable]
 public class SaveLoadSceneData 
 {
-    public int bestScore;
+    public float bestScore;
 }
